+++ lwk_wasm/tests/node/amp2.js
@@ -23,9 +23,9 @@ async function runAmp2Test() {
         // getting the descriptor with signer managed externally leads to the same descriptor
         const userPath = amp2.elip153UserPath(account);
         const viewPath = amp2.elip153ViewPath(account);
-        const userKeyoriginXpub = signer.keyoriginXpubFromPath(userPath);
-        const viewKeyoriginXpub = signer.keyoriginXpubFromPath(viewPath);
-        const descFromStr = amp2.elip153FromStr(userKeyoriginXpub, viewKeyoriginXpub);
+        const userKeyoriginXpub = signer.deriveXpub(userPath);
+        const viewKeyoriginXpub = signer.deriveXpub(viewPath);
+        const descFromStr = amp2.elip153FromExternalSigner(account, userKeyoriginXpub, viewKeyoriginXpub);
         assert.strictEqual(desc.descriptor().toString(), descFromStr.descriptor().toString());
 
         const wollet = new lwk.Wollet(network, desc.descriptor());