+++ lwk_wollet/src/amp2.rs
@@ -156,10 +156,10 @@ impl Amp2 {
     pub fn elip153_from_signer<S: Signer>(
         &self,
         signer: &S,
-        account: u32,
+        account_num: u32,
     ) -> Result<Amp2Descriptor, crate::Error> {
-        let user_path = self.elip153_user_path(account)?;
-        let view_path = self.elip153_view_path(account)?;
+        let user_path = self.elip153_user_path(account_num)?;
+        let view_path = self.elip153_view_path(account_num)?;
 
         // TODO: map signer errors more nicely
         let fingerprint = signer
@@ -181,12 +181,12 @@ impl Amp2 {
     }
 
     /// ELIP153 `USER_PATH = m/purpose'/coin_type'/account'`
-    pub fn elip153_user_path(&self, account: u32) -> Result<DerivationPath, crate::Error> {
+    pub fn elip153_user_path(&self, account_num: u32) -> Result<DerivationPath, crate::Error> {
         let coin_type = if self.is_mainnet { 1776 } else { 1 };
         Ok(DerivationPath::from(vec![
             ChildNumber::from_hardened_idx(ELIP153_PURPOSE)?,
             ChildNumber::from_hardened_idx(coin_type)?,
-            ChildNumber::from_hardened_idx(account)?,
+            ChildNumber::from_hardened_idx(account_num)?,
         ]))
     }
 
@@ -210,29 +210,47 @@ impl Amp2 {
     }
 
     /// ELIP153 `VIEW_PATH = m/purpose'/coin_type'/account'/server_fingerprint_masked'`
-    pub fn elip153_view_path(&self, account: u32) -> Result<DerivationPath, crate::Error> {
-        let user_path = self.elip153_user_path(account)?;
+    pub fn elip153_view_path(&self, account_num: u32) -> Result<DerivationPath, crate::Error> {
+        let user_path = self.elip153_user_path(account_num)?;
         let server_fingerprint_masked =
             u32::from_be_bytes(self.server_xpub.fingerprint().to_bytes()) & 0x7FFF_FFFF;
         Ok(user_path.child(ChildNumber::from_hardened_idx(server_fingerprint_masked)?))
     }
 
-    /// Create an AMP2 descriptor ELIP153 compliant from xpub strings.
+    /// Create an AMP2 descriptor ELIP153 compliant from a signer managed externally.
     ///
-    /// This is typically used when the signer is managed outside of LWK.
-    /// Derive the user xpub at [`Amp2::elip153_user_path()`] and
-    /// the view xpub at [`Amp2::elip153_view_path()`], and pass the
-    /// obtained keyorigin_xpub strings here.
-    pub fn elip153_from_str(
+    /// Caller must ensure that:
+    /// * `user_keyorigin_xpub` is the keyorigin xpub derived at [`Amp2::elip153_user_path()`] for `account_num`
+    /// * `view_keyorigin_xpub` is the keyorigin xpub derived at [`Amp2::elip153_view_path()`] for `account_num`
+    ///
+    /// **Warning**: Passing incorrect signer data can lead to creating an incorrect
+    /// descriptor, which could lead to loss of funds.
+    pub fn elip153_from_external_signer(
         &self,
-        user_keyorigin_xpub: &str,
-        view_keyorigin_xpub: &str,
+        account_num: u32,
+        user_keyorigin_xpub: (KeySource, Xpub),
+        view_keyorigin_xpub: (KeySource, Xpub),
     ) -> Result<Amp2Descriptor, crate::Error> {
-        let (user_keysource, user_xpub) = keyorigin_xpub_from_str(user_keyorigin_xpub)?;
-        let user_keysource = user_keysource.ok_or(crate::Error::MissingKeyorigin)?;
-        let (view_keysource, view_xpub) = keyorigin_xpub_from_str(view_keyorigin_xpub)?;
-        let view_keysource = view_keysource.ok_or(crate::Error::MissingKeyorigin)?;
-        self.elip153(user_keysource, user_xpub, view_keysource, view_xpub)
+        let ((user_fp, user_path), user_xpub) = user_keyorigin_xpub;
+        let ((view_fp, view_path), view_xpub) = view_keyorigin_xpub;
+
+        if user_fp != view_fp {
+            return Err(crate::Error::Generic("fingerprint mismatch".to_string()));
+        }
+        if user_path != self.elip153_user_path(account_num)?
+            || view_path != self.elip153_view_path(account_num)?
+        {
+            return Err(crate::Error::Generic(
+                "unexpected keyorigin derivation path".to_string(),
+            ));
+        }
+
+        self.elip153(
+            (user_fp, user_path),
+            user_xpub,
+            (view_fp, view_path),
+            view_xpub,
+        )
     }
 
     fn elip153(
@@ -461,7 +479,8 @@ mod test {
         use lwk_signer::SwSigner;
 
         let mnemonic = "abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon abandon about";
-        let signer = SwSigner::new(mnemonic, false).unwrap();
+        let signer =
+            SwSigner::new_with_network(mnemonic, lwk_common::Network::TestnetLiquid).unwrap();
 
         let amp2 = Amp2::new_testnet();
         let account = 0;
@@ -551,9 +570,8 @@ mod test {
             ),
         ] {
             i += 1;
-            let is_mainnet = network == Network::Liquid;
-            let signer = SwSigner::new(mnemonic, is_mainnet).unwrap();
-            let server_signer = SwSigner::new(server_mnemonic, is_mainnet).unwrap();
+            let signer = SwSigner::new_with_network(mnemonic, network).unwrap();
+            let server_signer = SwSigner::new_with_network(server_mnemonic, network).unwrap();
             let server_xpub = server_signer.xpub();
 
             let amp2 = Amp2 {
@@ -561,7 +579,7 @@ mod test {
                 server_xpub,
                 server_fingerprint: server_xpub.fingerprint(),
                 url: URL_TESTNET.into(),
-                is_mainnet,
+                is_mainnet: network.is_mainnet(),
             };
             let (user_keysource, user_xpub) =
                 derive(&signer, &amp2.elip153_user_path(account).unwrap());