+++ lwk_bindings/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Added `DescriptorPublicKey`
+* Added `Signer::ss_desc()`
+* Renamed `DerivationPath::from_account()` to `DerivationPath::ss_path()`
+* Replaced `WolletDescriptor::from_xpub()` with `WolletDescriptor::ss_desc_from_external_signer()`
+* Replaced `Amp2::elip153_from_str()` with `Amp2::elip153_from_external_signer()`
+* Replaced `Signer::keyorigin_xpub_from_path()` with `Signer::derive_xpub()`
+
 ## 0.19.0
 
 * Added `DerivationPath`